requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.7.7
anyio==3.7.1
attrs==25.3.0
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
fastapi==0.104.1
filelock==3.18.0
flake8==6.1.0
frozenlist==1.6.2
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.32.4
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.4.4
mypy_extensions==1.1.0
openai==1.3.7
packaging==25.0
pathspec==0.12.1
Pillow==10.4.0
platformdirs==4.3.8
pluggy
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.11.5
pydantic-settings==2.1.0
pydantic_core==2.33.2
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv>=1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tokenizers==0.21.1
tqdm==4.67.1
twilio==8.10.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1